refactor: make PendingStats template-based for flexibility
