+++ src/DRAMPower/CMakeLists.txt
@@ -29,7 +29,6 @@ add_library(DRAMPower
     DRAMPower/util/binary_ops.cpp
     DRAMPower/util/extensions.cpp
     DRAMPower/util/pin.cpp
-    DRAMPower/util/pending_stats.cpp
 )
 
 target_include_directories(DRAMPower PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
